style: add graphic to (unauthenticated) home page
